Add files via upload
added new testdata as JSON in the specified file format. File is containing the following attributes: ID (= PtNumber in the existing files), coords in WGS84 coordinate system (EPSG code 4326), point velocity for coloring the points with a certain color scale (velocity is actually random, but this should be changed) and the data points to visualize in a second plot.